requirements.txt
chromadb==0.5.2
pdfplumber==0.11.1
protobuf==5.27.1
python-dotenv==1.0.1
google-generativeai==0.7.0